requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
celery==5.6.2
redis>=5.0.0
pandas
openpyxl
python-multipart